backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-frontmatter==1.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
